% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_pop_centroids.R
\name{county_pop_centroids}
\alias{county_pop_centroids}
\title{County Population Centers}
\usage{
county_pop_centroids(year = 2020)
}
\arguments{
\item{year}{integer, year of data release}
}
\value{
A \code{data.table} containing:
\itemize{
\item \code{geoid}: concatenated state and county FIPS code
\item \code{lon}: longitude of county population center
\item \code{lat}: latitude of county population center
}
}
\description{
This function downloads the 2020 U.S. county population center data from the
Census Bureau, constructs a combined GEOID, and returns a table with
\code{geoid}, \code{lon}, and \code{lat}.
}
\examples{

\dontrun{
 county_pop_centers <- county_pop_centroids(2020)
 head(county_centers)
}

}
