% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_from_url.R
\name{download_from_url}
\alias{download_from_url}
\title{Download a file from a URL}
\usage{
download_from_url(url, file_name, output_dir = ".")
}
\arguments{
\item{url}{URL to download from}

\item{file_name}{Name for the file}

\item{output_dir}{Directory for output. Defaults to current working directory}
}
\value{
The file path, to be passed on to other functions
}
\description{
Download a file from a URL
}
