% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geojson_array_to_sf}
\alias{geojson_array_to_sf}
\title{A function that can turn an array of GeoJSON objects into an sf data.fame}
\usage{
geojson_array_to_sf(geo_json_list)
}
\arguments{
\item{geo_json_list}{A list of GeoJSON objects (i.e. the "features" array property of a FeatureCollection)}
}
\value{
sf data.frame
}
\description{
A function that can turn an array of GeoJSON objects into an sf data.fame
}
