% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drive_time_polygons.R
\name{get_drive_time_polygons}
\alias{get_drive_time_polygons}
\title{Get drive time polygons around points}
\usage{
get_drive_time_polygons(
  points,
  column_lon,
  column_lat,
  drivetime,
  res = 30,
  max_attempt = 10
)
}
\arguments{
\item{points}{A data.frame of point data}

\item{column_lon}{Column name of longitude column}

\item{column_lat}{Column name of latitude column}

\item{drivetime}{Length of drive time in minutes (e.g. 30 for 30 minutes)}

\item{res}{Resolution. Higher resolution yields more detailed polygons and takes more time.}

\item{max_attempt}{Maximum number of tries to retrueve polygons. Defaults to 10}
}
\value{
A spatial data frame with drive time polygons
}
\description{
Get drive time polygons around points
}
