% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris_helper.R
\name{load_tiger}
\alias{load_tiger}
\title{(Adapted from tigris) Helper function to download Census data (modified to read from local file paths)}
\usage{
load_tiger(
  local_file_path,
  tigris_type = NULL,
  class = getOption("tigris_class", "sf"),
  filter_by = NULL
)
}
\arguments{
\item{local_file_path}{local path to zipped shapefile in TIGER database (constructed in calling function).}

\item{tigris_type}{Added as an attribute to return object (used internally).}

\item{class}{Class of return object. Must be one of "sf" (the default) or "sp".}

\item{filter_by}{Geometry used to filter the output returned by the function.  Can be an sf object, an object of class \code{bbox}, or a length-4 vector of format \code{c(xmin, ymin, xmax, ymax)} that can be converted to a bbox. Geometries that intersect the input to \code{filter_by} will be returned.}
}
\value{
sf or sp data frame
}
\description{
(Adapted from tigris) Helper function to download Census data (modified to read from local file paths)
}
\examples{

\dontrun{
  states <- lapply(local_state_files_tiger_year, load_tiger) |>
    dplyr::bind_rows()
}

}
