% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs.R
\name{read_acs}
\alias{read_acs}
\title{Read and de-normalize ACS data from Postgres}
\usage{
read_acs(
  con,
  acs_table,
  variables,
  years = 2019,
  variables_from = "variable",
  values_from = "estimate"
)
}
\arguments{
\item{con}{A database connection to the ACS schema}

\item{acs_table}{Name of the table to read from}

\item{variables}{A vector of variable names to read}

\item{years}{A vector of years to pull data for}

\item{variables_from}{Where to pull column names from, one of 'variable' or 'acs_code'}

\item{values_from}{Where to pull data values from, one of 'estimate' or 'moe'}
}
\value{
A de-normalized table, unique on geoid and year
}
\description{
Read and de-normalize ACS data from Postgres
}
