% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_line_county_state_fips.R
\name{tiger_line_county_state_fips}
\alias{tiger_line_county_state_fips}
\title{Function to load data.frame that maps County FIPS codes to States}
\usage{
tiger_line_county_state_fips(tiger_year = 2024)
}
\arguments{
\item{tiger_year}{integer, year of data release}
}
\value{
return Census County & State FIPS codes as data.frame
}
\description{
Function to load data.frame that maps County FIPS codes to States
}
\examples{

\dontrun{
 county_state_fips <- tiger_line_county_state_fips(2024)
}

}
