% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_line_places.R
\name{tiger_line_places}
\alias{tiger_line_places}
\title{Function to load Census Place boundaries from S3 (or from local disk, if cached as .RDS)}
\usage{
tiger_line_places(tiger_year = 2024)
}
\arguments{
\item{tiger_year}{integer, year of data release}
}
\value{
return Census Place boundaries as sf data.frame
}
\description{
Function to load Census Place boundaries from S3 (or from local disk, if cached as .RDS)
}
\examples{

\dontrun{
 places <- tiger_line_places(2024)
}

}
