% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_line_states.R
\name{tiger_line_states}
\alias{tiger_line_states}
\title{Function to load Census State boundaries from S3 (or from local disk, if cached as .RDS)}
\usage{
tiger_line_states(year = "cb_2024")
}
\arguments{
\item{year}{character, year of data release; prefix with "cb_" for cartographic boundaries}
}
\value{
return Census State boundaries as sf data.frame object
}
\description{
Function to load Census State boundaries from S3 (or from local disk, if cached as .RDS)
}
\examples{

\dontrun{
 states <- tiger_line_states("2024") # Returns full tiger line 2024 state boundary geomtry
}

}
