% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_airtable_column.R
\name{update_airtable_column}
\alias{update_airtable_column}
\title{Update an existing airtable column from R}
\usage{
update_airtable_column(
  dta,
  record_id_col,
  data_col,
  api_key,
  base_id,
  table_name,
  variable_name,
  type = c("character", "numeric", "multi_select")
)
}
\arguments{
\item{dta}{A data frame}

\item{record_id_col}{Unquoted name of the column corresponding to the airtable record ID}

\item{data_col}{Unquoted name of the column to update}

\item{api_key}{Airtable API key}

\item{base_id}{The ID of the airtable base}

\item{table_name}{Name of the airtable table to update}

\item{variable_name}{Name of the variable to update}

\item{type}{Data type, one of character, numeric, or multi_select}
}
\description{
Update an existing airtable column from R
}
