% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_carto.R
\name{write_carto}
\alias{write_carto}
\title{Write spatial data to Carto}
\usage{
write_carto(
  sf,
  layer_name,
  carto_conf = "oscarto",
  overwrite = FALSE,
  config_group = "write_carto",
  config_file = "/data/Github/base/config.yml"
)
}
\arguments{
\item{sf}{A spatial dataframe with class 'sf'}

\item{layer_name}{The name of the data set as it should appear on Carto}

\item{carto_conf}{Name of the YAML header containing the URL to upload to. Defaults to oscarto}

\item{overwrite}{Option to overwrite existing data}

\item{config_group}{Name of the YAML collection containing upload URLs.}

\item{config_file}{Path to the config YAML with credentials}
}
\description{
Write spatial data to Carto
}
