% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_code_centroids.R
\name{zip_code_centroids}
\alias{zip_code_centroids}
\title{Function to load ZIP Code Population Weighted Centroids from S3 (or from local disk, if cached as .RDS)
https://catalog.data.gov/dataset/zip-code-population-weighted-centroids}
\usage{
zip_code_centroids(hud_year = 2024)
}
\arguments{
\item{hud_year}{integer, year of data release}
}
\value{
return ZIP Code Population Weighted Centroids as sf data.frame object
}
\description{
Function to load ZIP Code Population Weighted Centroids from S3 (or from local disk, if cached as .RDS)
https://catalog.data.gov/dataset/zip-code-population-weighted-centroids
}
\examples{

\dontrun{
 states <- zip_code_centroids(2024)
}

}
