% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airtable.R
\name{get_airtable}
\alias{get_airtable}
\title{Get table from Airtable}
\usage{
get_airtable(
  key,
  base = "https://api_airtable_com/v0/appD6byygamXxc3jx/",
  table,
  view = "All",
  max_records = 500,
  print = FALSE
)
}
\arguments{
\item{key}{API key}

\item{base}{Airtable base to connect to}

\item{table}{Airtable table name}

\item{view}{Airtable view to connect to. Defaults to 'All'}

\item{max_records}{Maximum number of records to return}

\item{print}{Optionally print diagnostics along the way. Defaults to FALSE}
}
\value{
A data.frame
}
\description{
Get table from Airtable
}
