% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{map_locations_to_counties}
\alias{map_locations_to_counties}
\title{Function to get place or ZIP code centroids given a data.frame with either \code{location} or \code{postal_code}}
\usage{
map_locations_to_counties(dta_loc_or_postal)
}
\arguments{
\item{dta_loc_or_postal}{data.frame, which includes either a
\code{location} attributes of the form "City(Town), ST" or a
\code{postal_code} atttribute with a 5-digit ZIP code value}
}
\value{
return "data.frame" object with county geoid and point geometry based on Census place or ZIP code centroid
}
\description{
Function to get place or ZIP code centroids given a data.frame with either \code{location} or \code{postal_code}
}
\examples{

\dontrun{
 dta_loc_or_postal <- data.frame(
   rank = c(1),
   link =  c("Conduit Coders 5/6 profile"),
   name = c("Conduit Coders 5/6"),
   location = c("Ada, OK"),
   score = c(79))
 dta_centroid_and_geoid_co <- map_locations_to_counties(dta_loc_or_postal)
}

}
