% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_line_counties.R
\name{tiger_line_counties}
\alias{tiger_line_counties}
\title{Function to load Census County boundaries from S3 (or from local disk, if cached as .RDS)}
\usage{
tiger_line_counties(tiger_year = 2024)
}
\arguments{
\item{tiger_year}{integer, year of data release}
}
\value{
return Census County boundaries as sf data.frame
}
\description{
Function to load Census County boundaries from S3 (or from local disk, if cached as .RDS)
}
\examples{

\dontrun{
 counties <- tiger_line_counties(2024)
}

}
