% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_line_states.R
\name{tiger_line_states}
\alias{tiger_line_states}
\title{Function to load Census State boundaries from S3 (or from local disk, if cached as .RDS)}
\usage{
tiger_line_states(tiger_year = 2024)
}
\arguments{
\item{tiger_year}{integer, year of data release}
}
\value{
return Census State boundaries as "sf" "data.frame" object
}
\description{
Function to load Census State boundaries from S3 (or from local disk, if cached as .RDS)
}
\examples{

\dontrun{
 states <- tiger_line_states(2024)
}

}
